feat: implement PreviewMedia component for handling media previews with fallback options
