+++ src/components/UI/Form/EmojiInput/Editor.tsx
@@ -37,9 +37,6 @@ export interface EditorProps {
   inputProp?: any;
   onChange?: any;
   defaultValue?: any;
-  // squares off this box's bottom corners/border so an attached section
-  // rendered directly below it (see TemplateVariables' `attached` prop) reads
-  // as a continuation of the same element instead of a separate floating box.
   squareBottom?: boolean;
 }
